start conda.common refactor
