+++ conda/cli/common.py
@@ -1,16 +1,16 @@
 from __future__ import print_function, division, absolute_import
 
-import re
-import os
-import sys
 import argparse
 import contextlib
-from os.path import abspath, basename, expanduser, isdir, join
+import os
+import re
+import sys
 import textwrap
+from os.path import abspath, basename, expanduser, isdir, join
 
 import conda.config as config
 from conda import console
-from conda.utils import memoize
+from conda.common.utils import memoize
 
 
 class Completer(object):
@@ -490,7 +490,7 @@ def spec_from_line(line):
 
 
 def specs_from_url(url, json=False):
-    from conda.fetch import TmpDownload
+    from conda.common.download import TmpDownload
 
     explicit = False
     with TmpDownload(url, verbose=False) as path:
