+++ conda/cli/main_remove.py
@@ -6,11 +6,10 @@
 
 from __future__ import print_function, division, absolute_import
 
-from os.path import join, exists
-
-from argparse import RawDescriptionHelpFormatter
 import errno
 import logging
+from argparse import RawDescriptionHelpFormatter
+from os.path import join, exists
 
 from conda.cli import common
 from conda.console import json_progress_bars
@@ -111,7 +110,7 @@ def execute(args, parser):
     channel_urls = args.channel or ()
     if args.use_local:
         from conda.fetch import fetch_index
-        from conda.utils import url_path
+        from conda.common.utils import url_path
         try:
             from conda_build.config import croot
         except ImportError:
