+++ conda/common/__init__.py
@@ -1,0 +1,3 @@
+# What often gets called "utils" or "misc."
+# All code must be able to stand independently on its own. (Think `@staticmethod`.)
+# Does NOT import from conda modules other than those in conda.common.
