+++ conda/common/compat.py
