+++ conda/connection.py
@@ -6,22 +6,22 @@
 
 from __future__ import print_function, division, absolute_import
 
-from logging import getLogger
-import re
+import base64
+import cgi
+import email
+import ftplib
 import mimetypes
 import os
-import email
-import base64
-import ftplib
-import cgi
-from io import BytesIO
+import re
 import tempfile
-
-import conda
-from conda.compat import urlparse, StringIO
-from conda.config import get_proxy_servers, ssl_verify
+from io import BytesIO
+from logging import getLogger
 
 import requests
+
+import conda
+from conda.common.compat import urlparse, StringIO
+from conda.config import get_proxy_servers, ssl_verify
 
 RETRIES = 3
