+++ conda/install.py
@@ -42,7 +42,7 @@ import traceback
 from os.path import abspath, basename, dirname, isdir, isfile, islink, join, relpath
 
 try:
-    from conda.lock import Locked
+    from conda.common.lock import Locked
 except ImportError:
     # Make sure this still works as a standalone script for the Anaconda
     # installer.
