+++ conda/plan.py
@@ -10,19 +10,19 @@ NOTE:
 
 from __future__ import print_function, division, absolute_import
 
-import sys
 import os
-from logging import getLogger
+import sys
 from collections import defaultdict
+from logging import getLogger
 from os.path import abspath, basename, isfile, join, exists
 
 from conda import config
 from conda import install
+from conda import instructions as inst
+from conda.common.utils import md5_file, human_bytes
+from conda.exceptions import CondaException
 from conda.history import History
 from conda.resolve import MatchSpec, Resolve, Package
-from conda.utils import md5_file, human_bytes
-from conda import instructions as inst
-from conda.exceptions import CondaException
 log = getLogger(__name__)
 
 # For backwards compatibility
