+++ conda/resolve.py
@@ -4,12 +4,12 @@ import logging
 from collections import defaultdict
 from itertools import chain
 
-from conda.compat import iterkeys, itervalues, iteritems, string_types
-from conda.logic import minimal_unsatisfiable_subset, Clauses
-from conda.version import VersionSpec, normalized_version
-from conda.console import setup_handlers
 from conda import config
+from conda.common.compat import iterkeys, itervalues, iteritems, string_types
+from conda.console import setup_handlers
+from conda.logic import minimal_unsatisfiable_subset, Clauses
 from conda.toposort import toposort
+from conda.version import VersionSpec, normalized_version
 
 log = logging.getLogger(__name__)
 dotlog = logging.getLogger('dotupdate')
